Combine 2D functionality in mixin
